@@ -1,17 +1,16 @@
-import { BrowserModule } from "@angular/platform-browser";
-import { NgModule } from "@angular/core";
-import { FormsModule } from "@angular/forms";
-import { HttpClientModule } from "@angular/common/http";
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { HttpClientModule, HttpClient } from '@angular/common/http';
 
-import { AppComponent } from "./app.component";
-
+import { AppComponent } from './app.component';
 // Custom Components
-import { SearchBarComponent } from "../../../app/src/app/components/search-bar/search-bar.component";
+import { SearchBarComponent } from './components/search-bar/search-bar.component';
+import { VenueItemComponent } from './components/venue-item/venue-item.component';
+import { SpinnerComponent } from './components/spinner/spinner.component';
 
 // Custom Services
-import { SearchService } from "./services/search.service";
-import { VenueItemComponent } from "../../../app/src/app/components/venue-item/venue-item.component";
-import { SpinnerComponent } from "../../../app/src/app/components/spinner/spinner.component";
+import { SearchService } from './services/search.service';
 
 @NgModule({
   declarations: [
