@@ -1,44 +1,35 @@
-import { TestBed, async } from "@angular/core/testing";
-import { FormsModule } from "@angular/forms";
-import { AppComponent } from "./app.component";
+import { TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
+import { AppComponent } from './app.component';
 
-// Components
-import { SearchBarComponent } from "../../../app/src/app/components/search-bar/search-bar.component";
-import { VenueItemComponent } from "../../../app/src/app/components/venue-item/venue-item.component";
-import { SpinnerComponent } from "../../../app/src/app/components/spinner/spinner.component";
-
-import { AbstractMockObservableService } from "./services/mock.service";
-import { SearchService } from "./services/search.service";
-import { HttpClient, HttpClientModule } from "@angular/common/http";
-
-class MockService extends AbstractMockObservableService {
-  doStuff() {
-    return this;
-  }
-}
-let mockService;
-describe("AppComponent", () => {
-  beforeEach(async(() => {
-    mockService = new MockService();
-    TestBed.configureTestingModule({
-      imports: [FormsModule],
-      declarations: [
-        AppComponent,
-        SearchBarComponent,
-        VenueItemComponent,
-        SpinnerComponent,
+describe('AppComponent', () => {
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [
+        RouterTestingModule
       ],
-      providers: [
-        { provide: SearchService, useValue: mockService },
-        HttpClientModule,
-        HttpClient,
+      declarations: [
+        AppComponent
       ],
     }).compileComponents();
-  }));
+  });
 
-  it("should create the app", async(() => {
+  it('should create the app', () => {
     const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.debugElement.componentInstance;
+    const app = fixture.componentInstance;
     expect(app).toBeTruthy();
-  }));
+  });
+
+  it(`should have as title 'app'`, () => {
+    const fixture = TestBed.createComponent(AppComponent);
+    const app = fixture.componentInstance;
+    expect(app.title).toEqual('app');
+  });
+
+  it('should render title', () => {
+    const fixture = TestBed.createComponent(AppComponent);
+    fixture.detectChanges();
+    const compiled = fixture.nativeElement;
+    expect(compiled.querySelector('.content span').textContent).toContain('app app is running!');
+  });
 });
